Adjusted `marlowe-playground-server` to API changes
