+++ marlowe-actus/actus-core/src/Actus/Domain.hs
@@ -1,3 +1,6 @@
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE RecordWildCards   #-}
 module Actus.Domain
   ( module Actus.Domain.BusinessEvents
   , module Actus.Domain.ContractState
@@ -9,6 +12,7 @@ module Actus.Domain
   , ScheduleOps (..)
   , RoleSignOps (..)
   , marloweFixedPoint
+  , setDefaultContractTermValues
   )
   where
 
@@ -16,6 +20,9 @@ import Actus.Domain.BusinessEvents
 import Actus.Domain.ContractState
 import Actus.Domain.ContractTerms
 import Actus.Domain.Schedule
+import Actus.Utility.YearFraction (yearFraction)
+import Control.Applicative ((<|>))
+import Control.Monad (guard)
 import Data.Time (LocalTime)
 
 marloweFixedPoint :: Integer
@@ -60,3 +67,70 @@ class (ActusNum a, ActusOps a) => RoleSignOps a where
     _r CR_PFL = _negate _one
     _r CR_RF  = _one
     _r CR_PF  = _negate _one
+
+-- == Default instance (Double)
+
+setDefaultContractTermValues :: ContractTerms Double -> ContractTerms Double
+setDefaultContractTermValues ct@ContractTerms {..} =
+  ct
+    { scheduleConfig =
+        scheduleConfig
+          { endOfMonthConvention = applyDefault EOMC_SD (endOfMonthConvention scheduleConfig),
+            businessDayConvention = applyDefault BDC_NULL (businessDayConvention scheduleConfig),
+            calendar = applyDefault CLDR_NC (calendar scheduleConfig)
+          },
+      contractPerformance            = applyDefault PRF_PF contractPerformance,
+      interestCalculationBase        = applyDefault IPCB_NT interestCalculationBase,
+      premiumDiscountAtIED           = applyDefault 0.0 premiumDiscountAtIED,
+      scalingEffect                  = applyDefault SE_OOO scalingEffect,
+      penaltyRate                    = applyDefault 0.0 penaltyRate,
+      penaltyType                    = applyDefault PYTP_O penaltyType,
+      prepaymentEffect               = applyDefault PPEF_N prepaymentEffect,
+      rateSpread                     = applyDefault 0.0 rateSpread,
+      rateMultiplier                 = applyDefault 1.0 rateMultiplier,
+      feeAccrued                     = applyDefault 0.0 feeAccrued,
+      feeRate                        = applyDefault 0.0 feeRate,
+      accruedInterest                = applyDefault 0.0 accruedInterest,
+      nominalInterestRate            = applyDefault 0.0 nominalInterestRate,
+      priceAtPurchaseDate            = applyDefault 0.0 priceAtPurchaseDate,
+      priceAtTerminationDate         = applyDefault 0.0 priceAtTerminationDate,
+      scalingIndexAtContractDealDate = applyDefault 0.0 scalingIndexAtContractDealDate,
+      periodFloor                    = applyDefault (- infinity) periodFloor,
+      periodCap                      = applyDefault infinity periodCap,
+      lifeCap                        = applyDefault infinity lifeCap,
+      lifeFloor                      = applyDefault (- infinity) lifeFloor,
+      interestCalculationBaseA       = applyDefault 0.0 interestCalculationBaseA,
+
+      -- see ContractModel.java
+      cycleAnchorDateOfInterestPayment = cycleAnchorDateOfInterestPayment <|> ((guard $ contractType == CLM) >> initialExchangeDate)
+    }
+  where
+    infinity :: Double
+    infinity = 1 Prelude./ 0 :: Double
+
+    applyDefault :: a -> Maybe a -> Maybe a
+    applyDefault v o = o <|> Just v
+
+
+instance RoleSignOps Double
+
+instance ActusOps Double where
+    _min  = min
+    _max  = max
+    _abs  = abs
+    _zero = 0.0
+    _one  = 1.0
+    _fromInteger = fromInteger
+    _negate = negate
+
+instance ActusNum Double where
+    a + b       = a Prelude.+ b
+    a - b       = a Prelude.- b
+    a * b       = a Prelude.* b
+    a / b       = a Prelude./ b
+
+instance YearFractionOps Double where
+    _y = yearFraction
+
+instance ScheduleOps Double where
+    _ceiling = ceiling
