+++ marlowe-actus/actus-core/src/Actus/Domain/ContractTerms.hs
@@ -8,10 +8,11 @@
 
 module Actus.Domain.ContractTerms where
 
+import Control.Applicative ((<|>))
 import Control.Monad (guard, mzero)
 import Data.Aeson.TH (deriveJSON)
-import Data.Aeson.Types (FromJSON, Options (..), Parser, ToJSON, Value (..), defaultOptions, object, parseJSON, toJSON,
-                         (.=))
+import Data.Aeson.Types (FromJSON, Options (..), Parser, SumEncoding (..), ToJSON, Value (..), defaultOptions,
+                         genericParseJSON, object, parseJSON, toJSON, (.:), (.:?), (.=))
 import Data.Maybe (fromMaybe)
 import Data.Text as T hiding (reverse, takeWhile)
 import Data.Text.Read as T
@@ -482,3 +483,102 @@ data ContractTerms a = ContractTerms
   }
   deriving stock (Show, Generic)
   deriving anyclass (ToJSON)
+
+instance FromJSON (Reference Double) where
+  parseJSON = genericParseJSON defaultOptions { sumEncoding = UntaggedValue }
+
+instance FromJSON (ContractStructure Double) where
+  parseJSON (Object v) =
+    ContractStructure
+      <$> v .: "object"
+      <*> v .: "referenceType"
+      <*> v .: "referenceRole"
+  parseJSON _ = mzero
+
+instance FromJSON (ContractTerms Double) where
+  parseJSON (Object v) =
+    ContractTerms
+      <$> (v .:  "contractID" <|> v .: "contractId")
+      <*> v .:  "contractType"
+      <*> (v .: "contractStructure" <|> return [])
+      <*> v .:  "contractRole"
+      <*> v .:? "settlementCurrency"
+      <*> v .:? "initialExchangeDate"
+      <*> v .:? "dayCountConvention"
+      <*> (v .: "scheduleConfig"
+           <|> ScheduleConfig
+               <$> v .:? "calendar"
+               <*> v .:? "endOfMonthConvention"
+               <*> v .:? "businessDayConvention"
+           )
+      <*> v .:  "statusDate"
+      <*> v .:? "contractPerformance"
+      <*> v .:? "creditEventTypeCovered"
+      <*> v .!? "coverageOfCreditEnhancement"
+      <*> v .!? "guaranteedExposure"
+      <*> v .:? "cycleOfFee"
+      <*> v .:? "cycleAnchorDateOfFee"
+      <*> v .:? "feeAccrued"
+      <*> v .:? "feeBasis"
+      <*> v .!? "feeRate"
+      <*> v .:? "cycleAnchorDateOfInterestPayment"
+      <*> v .:? "cycleOfInterestPayment"
+      <*> v .!? "accruedInterest"
+      <*> v .:? "capitalizationEndDate"
+      <*> v .:? "cycleAnchorDateOfInterestCalculationBase"
+      <*> v .:? "cycleOfInterestCalculationBase"
+      <*> v .:? "interestCalculationBase"
+      <*> v .!? "interestCalculationBaseAmount"
+      <*> v .!? "nominalInterestRate"
+      <*> v .!? "nominalInterestRate2"
+      <*> v .!? "interestScalingMultiplier"
+      <*> v .:? "maturityDate"
+      <*> v .:? "amortizationDate"
+      <*> v .:? "exerciseDate"
+      <*> v .!? "notionalPrincipal"
+      <*> v .!? "premiumDiscountAtIED"
+      <*> v .:? "cycleAnchorDateOfPrincipalRedemption"
+      <*> v .:? "cycleOfPrincipalRedemption"
+      <*> v .!? "nextPrincipalRedemptionPayment"
+      <*> v .:? "purchaseDate"
+      <*> v .!? "priceAtPurchaseDate"
+      <*> v .:? "terminationDate"
+      <*> v .!? "priceAtTerminationDate"
+      <*> v .!? "quantity"
+      <*> v .:? "scalingIndexAtStatusDate"
+      <*> v .:? "cycleAnchorDateOfScalingIndex"
+      <*> v .:? "cycleOfScalingIndex"
+      <*> v .:? "scalingEffect"
+      <*> v .!? "scalingIndexAtContractDealDate"
+      <*> v .:? "marketObjectCodeOfScalingIndex"
+      <*> v .!? "notionalScalingMultiplier"
+      <*> v .:? "cycleOfOptionality"
+      <*> v .:? "cycleAnchorDateOfOptionality"
+      <*> v .:? "optionType"
+      <*> v .!? "optionStrike1"
+      <*> v .:? "optionExerciseType"
+      <*> v .:? "settlementPeriod"
+      <*> v .:? "deliverySettlement"
+      <*> v .!? "exerciseAmount"
+      <*> v .!? "futuresPrice"
+      <*> v .:? "penaltyRate"
+      <*> v .:? "penaltyType"
+      <*> v .:? "prepaymentEffect"
+      <*> v .:? "cycleOfRateReset"
+      <*> v .:? "cycleAnchorDateOfRateReset"
+      <*> v .!? "nextResetRate"
+      <*> v .!? "rateSpread"
+      <*> v .!? "rateMultiplier"
+      <*> v .:? "periodFloor"
+      <*> v .:? "periodCap"
+      <*> v .:? "lifeCap"
+      <*> v .:? "lifeFloor"
+      <*> v .:? "marketObjectCodeOfRateReset"
+      <*> v .:? "cycleOfDividendPayment"
+      <*> v .:? "cycleAnchorDateOfDividendPayment"
+      <*> v .:? "nextDividendPaymentAmount"
+      <*> (fromMaybe False <$> (v .:? "enableSettlement"))
+      <*> v .:? "constraints"
+    where
+      (.!?) w s = w .:? s <|> (fmap read <$> w .:? s)
+  parseJSON _ = mzero
