+++ marlowe-actus/marlowe-actus.cabal
@@ -84,7 +84,6 @@ test-suite actus-core-test
   type: exitcode-stdio-1.0
   main-is: Spec.hs
   other-modules:
-    Spec.Actus
     Spec.TestFramework
   build-depends:
     actus-core,
