+++ marlowe-playground-server/src/API.hs
@@ -6,13 +6,12 @@
 
 module API where
 
+import Actus.Marlowe (CashFlow, ContractTerms)
 import qualified Auth
 import Data.Aeson (FromJSON, ToJSON, Value)
 import Data.Text (Text)
 import GHC.Generics (Generic)
 import Language.Haskell.Interpreter (InterpreterError, InterpreterResult)
-import qualified Language.Marlowe.ACTUS.Domain.ContractTerms as A (ContractTerms)
-import qualified Language.Marlowe.ACTUS.Domain.Schedule as A (CashFlow)
 import Servant.API (Capture, Get, Header, Headers, JSON, NoContent, PlainText, Post, Raw, ReqBody, StdMethod (GET),
                     Verb, (:<|>), (:>))
 import Web.Cookie (SetCookie)
@@ -20,8 +19,7 @@ import Webghc.Server (CompileRequest)
 
 type API
      = "oracle" :> Capture "exchange" String :> Capture "pair" String :> Get '[JSON] Value
-       :<|> "actus" :> ("generate" :> ReqBody '[ JSON] A.ContractTerms :> Post '[ JSON] String
-                        :<|> "generate-static" :> ReqBody '[ JSON] A.ContractTerms :> Post '[ JSON] String
-                        :<|> "cashflows" :> ReqBody '[ JSON] A.ContractTerms :> Post '[ JSON] [A.CashFlow])
+       :<|> "actus" :> ("generate" :> ReqBody '[JSON] (ContractTerms Double) :> Post '[JSON] String
+                   :<|> "cashflows" :> ReqBody '[JSON] (ContractTerms Double) :> Post '[JSON] [CashFlow Double])
        :<|> "compile" :> ReqBody '[JSON] CompileRequest :> Post '[JSON] (Either InterpreterError (InterpreterResult String))
        :<|> "logout" :> Get '[JSON] (Headers '[Header "Set-Cookie" SetCookie] Value)
